+++ tests/apps/mip/test_apps_mip_files.py
@@ -2,7 +2,7 @@
 
 from snapshottest import Snapshot
 
-from cg.apps.mip_rna import files
+from cg.apps.mip import parse_sampleinfo
 
 
 def test_parse_sampleinfo_rna_result_contents(snapshot: Snapshot, files_raw: dict):
@@ -19,7 +19,7 @@ def test_parse_sampleinfo_rna_result_contents(snapshot: Snapshot, files_raw: dic
     sampleinfo_raw = files_raw["rna_sampleinfo"]
 
     # WHEN parsing the file
-    sampleinfo_data = files.parse_sampleinfo_rna(sampleinfo_raw)
+    sampleinfo_data = parse_sampleinfo.parse_sampleinfo_rna(sampleinfo_raw)
 
     # THEN the result should contains the data as specified by the MIP team
     snapshot.assert_match(sampleinfo_data)
